+++ tests/testthat.R
@@ -1,0 +1,4 @@
+library(testthat)
+library(BIGf90)
+
+test_check("BIGf90")